Basic Math Operations
This code demonstrates basic math operations in JavaScript using both let and var variable declarations.

The main difference between let and var variables is that with var, we can redeclare the same variables (ex: var Num2 = 2 is redeclared as var Num2 = 3). In the output, the value of Num2 is 3. But, with let we cannot declare variables with the same name. It is best to use let in coding because it is better than var for preventing errors.